# Web/super revenge/src/mission-control-system/docs/swagger.yaml
basePath: /api/v1
definitions:
  dtos.CreateCommentDto:
    properties:
      content:
        type: string
      noteId:
        type: integer
    type: object
...
  dtos.DeleteCommentDto:
    properties:
      id:
        type: integer
    type: object
...
  dtos.StructuredResponse:
    description: Standard response format containing success status, HTTP status code,
      message, and optional payload
    properties:
      message:
        description: |-
          Human-readable message
          @example Operation completed successfully
        example: Operation completed successfully
        type: string
      payload:
        description: |-
          Optional payload containing response data
          @example {"id":1,"name":"Example Item"}
      status:
        description: |-
          HTTP status code
          @example 200
        example: 200
        type: integer
      success:
        description: |-
          Whether the operation was successful
          @example true
        example: true
        type: boolean
    type: object
...
  dtos.UpdateCommentDto:
    properties:
      content:
        type: string
      id:
        type: integer
    type: object
...
host: localhost:3002
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A RESTful API for Todo management built with Go and PostgreSQL
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate Beginner Project
  version: "1.0"
paths:
...
  /comments/delete:
    delete:
      consumes:
      - application/json
      description: Delete an existing comment
      parameters:
      - description: Comment deletion data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dtos.DeleteCommentDto'
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
  /comments/edit:
    put:
      consumes:
      - application/json
      description: Update an existing comment
      parameters:
      - description: Comment update data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateCommentDto'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - comments
  /comments/new:
    post:
      consumes:
      - application/json
      description: Create a new comment for a note
      parameters:
      - description: Comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCommentDto'
      produces:
      - application/json
      responses:
        "200":
          description: Comment created successfully
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - comments
  /comments/note/{noteId}:
    get:
      consumes:
      - application/json
      description: Retrieve all comments belonging to the specified note
      parameters:
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comments retrieved successfully
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "400":
          description: Invalid note ID
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Get comments for a note
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"